add `no-std` to `sov-schema-db`
