+++ Cargo.toml
@@ -76,7 +76,7 @@ bytes = { version = "1.2.1", default-features = false }
 digest = { version = "0.10.6", default-features = false, features = ["alloc"] }
 futures = "0.3"
 hex = { version = "0.4.3", default-features = false, features = ["alloc", "serde"] }
-once_cell = "1.10.0"
+once_cell = { version = "1.10.0", default-features = false, features = ["alloc"] }
 prometheus = { version = "0.13.3", default-features = false }
 proptest = { version = "1.3.1", default-features = false, features = ["alloc"] }
 proptest-derive = "0.3.0"
@@ -89,7 +89,7 @@ sha2 = { version = "0.10.6", default-features = false }
 spin = "0.9.8"
 thiserror = "1.0.38"
 tiny-keccak = "2.0.2"
-tracing = "0.1.37"
+tracing = { version = "0.1.37", default-features = false }
 bech32 = "0.9.1"
 derive_more = "0.99.11"
 clap = { version = "4.2.7", features = ["derive"] }
