+++ full-node/db/sov-schema-db/Cargo.toml
@@ -14,13 +14,35 @@ readme = "README.md"
 
 [dependencies]
 # External dependencies 
-anyhow = { workspace = true, default-features = true }
-once_cell = { workspace = true }
-prometheus = { workspace = true }
-rocksdb = { workspace = true }
+anyhow = { workspace = true }
+once_cell = { workspace = true, optional = true, default-features = true }
+prometheus = { workspace = true, optional = true }
+rocksdb = { workspace = true, optional = true }
 tracing = { workspace = true }
-thiserror = { workspace = true }
+thiserror = { workspace = true, optional = true }
+
+sov-rollup-interface = { path = "../../../rollup-interface", default-features = false }
 
 [dev-dependencies]
 byteorder = { workspace = true }
 tempfile = { workspace = true }
+
+[features]
+default = ["std"]
+std = [
+    "anyhow/default",
+    "once_cell",
+    "prometheus",
+    "rocksdb",
+    "sov-rollup-interface/default",
+    "thiserror",
+    "tracing/default",
+]
+
+[[test]]
+name = "db_test"
+required-features = ["std"]
+
+[[test]]
+name = "iterator_test"
+required-features = ["std"]
